backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
aiosmtplib==3.0.1
python-dotenv==1.0.0
jinja2==3.1.3
weasyprint==60.2
